// src/modules/mgmt-api-client/model/contractDefinitionRequestDto.ts
/**
 * EDC REST API
 * EDC REST APIs - merged by OpenApiMerger
 *
 * The version of the OpenAPI document: 0.0.1-SNAPSHOT
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CriterionDto } from './criterionDto';


export interface ContractDefinitionRequestDto { 
    accessPolicyId: string;
    contractPolicyId: string;
    criteria: Array<CriterionDto>;
    id?: string;
    validity?: number;
}